+++ sources/calendar-1.x/src/Aeon/Calendar/Gregorian/Month.php
@@ -13,21 +13,26 @@ use Aeon\Calendar\TimeUnit;
  */
 final class Month
 {
+    private const TOTAL_MONTHS = 12;
+
     private Year $year;
 
     private MonthDays $days;
 
     private int $number;
 
+    private ?\DateTimeImmutable $dateTime;
+
     public function __construct(Year $year, int $number)
     {
-        if ($number <= 0 || $number > 12) {
+        if ($number <= 0 || $number > self::TOTAL_MONTHS) {
             throw new InvalidArgumentException('Month number must be greater or equal 1 and less or equal than 12');
         }
 
         $this->year = $year;
         $this->number = $number;
         $this->days = new MonthDays($this);
+        $this->dateTime = null;
     }
 
     /**
@@ -109,6 +114,7 @@ final class Month
     {
         $this->year = $data['year'];
         $this->number = $data['number'];
+        $this->dateTime = null;
         $this->days = new MonthDays($this);
     }
 
@@ -119,42 +125,93 @@ final class Month
 
     public function previous() : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify('-1 month'));
+        return $this->minusMonths(1);
     }
 
     public function next() : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify('+1 month'));
+        return $this->plusMonths(1);
     }
 
     public function plus(int $years, int $months) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('+%d months +%d years', $months, $years)));
+        $month = $this;
+
+        if ($months !== 0) {
+            $month = $month->plusMonths($months);
+        }
+
+        if ($years !== 0) {
+            $month = $month->plusYears($years);
+        }
+
+        return $month;
     }
 
     public function minus(int $years, int $months) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('-%d months -%d years', $months, $years)));
+        $month = $this;
+
+        if ($months !== 0) {
+            $month = $month->minusMonths($months);
+        }
+
+        if ($years !== 0) {
+            $month = $month->minusYears($years);
+        }
+
+        return $month;
     }
 
     public function plusMonths(int $months) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('+%d month', $months)));
+        $years = (int) ($months / self::TOTAL_MONTHS);
+        $monthsRemainder = $months % self::TOTAL_MONTHS;
+
+        $year = $this->year->number() + $years;
+        $month = $this->number() + $monthsRemainder;
+
+        if ($month > self::TOTAL_MONTHS) {
+            $year += ((int) ($month / self::TOTAL_MONTHS));
+            $month %= self::TOTAL_MONTHS;
+        }
+
+        if ($month === 0) {
+            $month = 12;
+        }
+
+        return new self(new Year($year), $month);
     }
 
     public function minusMonths(int $months) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('-%d month', $months)));
+        $years = (int) ($months / self::TOTAL_MONTHS);
+        $monthsRemainder = $months % self::TOTAL_MONTHS;
+
+        $year = $this->year->number() - $years;
+        $month = $this->number() - $monthsRemainder;
+
+        if ($month === 0) {
+            $month = self::TOTAL_MONTHS;
+            $year--;
+        }
+
+        if ($month < 0) {
+            $month = self::TOTAL_MONTHS - \abs($month);
+            $year--;
+        }
+
+        return new self(new Year($year), $month);
     }
 
     public function plusYears(int $years) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('+%d years', $years)));
+        return new self($this->year->plus($years), $this->number);
     }
 
     public function minusYears(int $years) : self
     {
-        return self::fromDateTime($this->toDateTimeImmutable()->modify(\sprintf('-%d years', $years)));
+        return new self($this->year->minus($years), $this->number);
     }
 
     public function firstDay() : Day
@@ -184,7 +241,8 @@ final class Month
 
     public function numberOfDays() : int
     {
-        return (int) $this->toDateTimeImmutable()->format('t');
+        return 31 - (($this->number == 2) ?
+                (3 - (int) $this->year->isLeap()) : (($this->number - 1) % 7 % 2));
     }
 
     public function shortName() : string
@@ -197,9 +255,18 @@ final class Month
         return $this->toDateTimeImmutable()->format('F');
     }
 
+    /**
+     * @psalm-suppress NullableReturnStatement
+     * @psalm-suppress InaccessibleProperty
+     * @psalm-suppress InvalidNullableReturnType
+     */
     public function toDateTimeImmutable() : \DateTimeImmutable
     {
-        return new \DateTimeImmutable(\sprintf('%d-%d-01 00:00:00.000000 UTC', $this->year()->number(), $this->number()));
+        if ($this->dateTime === null) {
+            $this->dateTime = new \DateTimeImmutable(\sprintf('%d-%d-01 00:00:00.000000 UTC', $this->year()->number(), $this->number()));
+        }
+
+        return $this->dateTime;
     }
 
     public function iterate(self $destination, Interval $interval) : Months
@@ -223,18 +290,12 @@ final class Month
             );
         }
 
-        return new Months(
-            ...\array_map(
-                function (\DateTimeImmutable $dateTimeImmutable) : self {
-                    return self::fromDateTime($dateTimeImmutable);
-                },
-                \iterator_to_array(
-                    $interval->toDatePeriod(
-                        $this->firstDay()->midnight(TimeZone::UTC()),
-                        RelativeTimeUnit::month(),
-                        $month->firstDay()->midnight(TimeZone::UTC())
-                    )
-                )
+        return Months::fromDateTimeIterator(
+            new DateTimeIntervalIterator(
+                $this->firstDay()->midnight(TimeZone::UTC()),
+                $month->firstDay()->midnight(TimeZone::UTC()),
+                RelativeTimeUnit::month(),
+                $interval
             )
         );
     }
@@ -253,20 +314,12 @@ final class Month
             );
         }
 
-        return new Months(
-            ...\array_map(
-                function (\DateTimeImmutable $dateTimeImmutable) : self {
-                    return self::fromDateTime($dateTimeImmutable);
-                },
-                \array_reverse(
-                    \iterator_to_array(
-                        $interval->toDatePeriodBackward(
-                            $month->firstDay()->midnight(TimeZone::UTC()),
-                            RelativeTimeUnit::month(),
-                            $this->firstDay()->midnight(TimeZone::UTC())
-                        )
-                    )
-                )
+        return Months::fromDateTimeIterator(
+            new DateTimeIntervalIterator(
+                $month->firstDay()->midnight(TimeZone::UTC()),
+                $this->firstDay()->midnight(TimeZone::UTC()),
+                RelativeTimeUnit::month(),
+                $interval
             )
         );
     }