+++ sources/calendar-1.x/src/Aeon/Calendar/Unit.php
@@ -8,4 +8,12 @@ namespace Aeon\Calendar;
 interface Unit
 {
     public function toDateInterval() : \DateInterval;
+
+    public function invert() : self;
+
+    public function isNegative() : bool;
+
+    public function toNegative() : self;
+
+    public function toPositive() : self;
 }